Making all permissions similar at every role & permission page

+++ database/migrations/2020_11_10_095507_create_rent_periods_table.php
@@ -16,7 +16,7 @@ class CreateRentPeriodsTable extends Migration
         Schema::create('rent_periods', function (Blueprint $table) {
             $table->tinyIncrements('id');
             $table->string('name', 100)->unique();
-            $table->unsignedTinyInteger('number_days')->nullable()->default(1);
+            $table->unsignedSmallInteger('number_days')->nullable()->default(1);
             // $table->string('code', 100);
             $table->softDeletes();
         });
